pyproject.toml
[project]
name = "aemo-energy-dashboard"
version = "0.1.0"
description = "Real-time Australian Energy Market (AEMO) electricity price and generation dashboard"
authors = [
    {name = "David Leitch", email = "david.leitch@icloud.com"}
]
readme = "README.md"
requires-python = ">=3.10"
keywords = ["aemo", "energy", "electricity", "dashboard", "australia"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: End Users/Desktop",
    "Topic :: Scientific/Engineering",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "panel>=1.0.0",
    "panel-material-ui>=0.1.0",
    "param>=2.0.0",
    "holoviews>=1.18.0",
    "hvplot>=0.9.0",
    "bokeh>=3.0.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "twilio>=8.0.0",
    "pyarrow>=12.0.0",
    "fastparquet>=2023.0.0",
    "matplotlib>=3.7.0",
    "matplotx>=0.3.0",
    "openpyxl>=3.1.0",
    "beautifulsoup4>=4.12.0",
    # Station analysis search functionality
    "fuzzywuzzy>=0.18.0",
    "python-levenshtein>=0.21.0",
    # FastAPI and data service dependencies
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.23.0",
    "pydantic>=2.0.0",
    "psutil>=7.0.0",
    "duckdb>=1.3.2",
    "plotly>=6.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
    "pre-commit>=3.0.0",
]

[project.scripts]
aemo-spot-update = "aemo_dashboard.spot_prices.update_spot:main"
aemo-spot-display = "aemo_dashboard.spot_prices.display_spot:main"
aemo-gen-dashboard = "aemo_dashboard.generation.gen_dash:main"
aemo-gen-update = "aemo_dashboard.generation.update_generation:main"
aemo-transmission-update = "aemo_dashboard.transmission.update_transmission:main"
aemo-transmission-backfill = "aemo_dashboard.transmission.backfill_transmission:main"
aemo-combined-update = "aemo_dashboard.combined.update_all:main"
aemo-manage-duids = "aemo_dashboard.scripts.manage_duid_exceptions:main"

[project.urls]
Homepage = "https://github.com/davidleitch/aemo-energy-dashboard"
Repository = "https://github.com/davidleitch/aemo-energy-dashboard.git"
Issues = "https://github.com/davidleitch/aemo-energy-dashboard/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/aemo_dashboard"]

[tool.black]
line-length = 88
target-version = ['py310']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["aemo_dashboard"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers --strict-config"
testpaths = ["tests"]
pythonpath = ["src"]
